feat: Update rate limit for clients in middleware.go
This commit updates the rate limit for clients in the middleware.go file. The previous rate limit was set to 4 requests per second with a burst limit of 8 requests. The new rate limit is set to 20 requests per second with a burst limit of 20 requests. This change improves the performance and scalability of the application.