Features/redux (#14)
* chore: Refactor code and update dependencies for frontend and backend

* chore: Refactor logout functionality and update logout button in navbar-profile.tsx

* feat: Remove unused ContactsQueryResponse interface

The ContactsQueryResponse interface is no longer used in the ContactsList component. This commit removes the unused interface to improve code cleanliness and maintainability.

* feat: Add postSlice reducer for managing post state

This commit adds a new postSlice reducer to the store, allowing for the management of post-related state in the application. The reducer includes actions for resetting the post state, incrementing and decrementing the view count, changing the post text, setting the privacy value, and setting the privacy visibleTo array. This addition improves the code organization and maintainability of the application.

* feat: Add lazy loading for images in home page and profile page

* feat: Update rate limit for clients in middleware.go

This commit updates the rate limit for clients in the middleware.go file. The previous rate limit was set to 4 requests per second with a burst limit of 8 requests. The new rate limit is set to 20 requests per second with a burst limit of 20 requests. This change improves the performance and scalability of the application.

* Convert tanstack -> redux

* split redux api endpoints to separate files

* Add automatic refresh token logic

* create separate auth middleware

* chore: Refactor Dockerfile for frontend

* chore: Add authMiddleware to store middleware

* chore: Update npm dependency to latest stable version

* fix some bugs

* remove prints and consolelogs

* Correct import

* group invite rework

* Close group creation dialog after successful creation

* Make reattending to group event possible

* Makes sure that id is there all time

* removed login dispatch from reauth request

* chore: Add forbiddenResponse method to handle forbidden access

* chore: Refactor unauthorized response to use forbiddenResponse method

* chore: Remove unused imports and optimize navbar component

* chore: Optimize navbar component and remove unused imports

---------

Co-authored-by: Kristo Veber <kristoveber7@gmail.com>